style: rename unsafe keys helper
