+++ object/property-path/src/index.ts
@@ -3,4 +3,4 @@ export * from './get.js'
 export * from './parse.js'
 export * from './set.js'
 export * from './token/index.js'
-export * from './unsafe-keys.js'
+export * from './unsafeKeys.js'
