+++ object/property-path/src/set.ts
@@ -1,7 +1,7 @@
 import { PnpmError } from '@pnpm/error'
 
 import { parsePropertyPath } from './parse.js'
-import { rejectUnsafeKeys } from './unsafe-keys.js'
+import { rejectUnsafeKeys } from './unsafeKeys.js'
 
 type ObjectOrArray = Record<string | number, unknown> | unknown[]
